+++ include/arch/arm/arch/32/mode/machine/debug.h
@@ -25,6 +25,7 @@
 #ifndef __ASSEMBLER__
 #include <stdint.h>
 #include <arch/machine/registerset.h>
+#include <armv/debug.h>
 
 void debug_init(void) VISIBLE;
 
@@ -167,28 +168,139 @@ bool_t isDebugFault(word_t hsr_or_fsr);
  */
 seL4_Fault_t handleUserLevelDebugException(word_t fault_vaddr);
 
-/** These next two functions are part of some state flags.
- *
- * A bitfield of all currently enabled breakpoints for a thread is kept in that
- * thread's TCB. These two functions here set and unset the bits in that
- * bitfield.
- */
-static inline void setBreakpointUsedFlag(tcb_t *t, uint16_t bp_num)
-{
-    if (t != NULL) {
-        t->tcbArch.tcbContext.breakpointState.used_breakpoints_bf |= BIT(bp_num);
-    }
-}
-
-static inline void unsetBreakpointUsedFlag(tcb_t *t, uint16_t bp_num)
-{
-    if (t != NULL) {
-        t->tcbArch.tcbContext.breakpointState.used_breakpoints_bf &= ~BIT(bp_num);
-    }
-}
-
 #endif /* CONFIG_HARDWARE_DEBUG_API */
 
 #endif /* !__ASSEMBLER__ */
 
 #endif /* defined(CONFIG_DEBUG_BUILD) || defined (CONFIG_HARDWARE_DEBUG_API) */
+
+#ifdef ARM_BASE_CP14_SAVE_AND_RESTORE
+
+#define MAKE_P14(crn, crm, opc2) "p14, 0, %0, c" #crn ", c" #crm ", " #opc2
+#define MAKE_DBGBVR(num) MAKE_P14(0, num, 4)
+#define MAKE_DBGBCR(num) MAKE_P14(0, num, 5)
+#define MAKE_DBGWVR(num) MAKE_P14(0, num, 6)
+#define MAKE_DBGWCR(num) MAKE_P14(0, num, 7)
+#define MAKE_DBGXVR(num) MAKE_P14(1, num, 1)
+
+/** Generates read functions for the CP14 control and value registers.
+ */
+#define DEBUG_GENERATE_READ_FN(_name, _reg) \
+static word_t \
+_name(uint16_t bp_num) \
+{ \
+    word_t ret; \
+ \
+    switch (bp_num) { \
+    case 1: \
+        MRC(MAKE_ ## _reg(1), ret); \
+        return ret; \
+    case 2: \
+        MRC(MAKE_ ## _reg(2), ret); \
+        return ret; \
+    case 3: \
+        MRC(MAKE_ ## _reg(3), ret); \
+        return ret; \
+    case 4: \
+        MRC(MAKE_ ## _reg(4), ret); \
+        return ret; \
+    case 5: \
+        MRC(MAKE_ ## _reg(5), ret); \
+        return ret; \
+    case 6: \
+        MRC(MAKE_ ## _reg(6), ret); \
+        return ret; \
+    case 7: \
+        MRC(MAKE_ ## _reg(7), ret); \
+        return ret; \
+    case 8: \
+        MRC(MAKE_ ## _reg(8), ret); \
+        return ret; \
+    case 9: \
+        MRC(MAKE_ ## _reg(9), ret); \
+        return ret; \
+    case 10: \
+        MRC(MAKE_ ## _reg(10), ret); \
+        return ret; \
+    case 11: \
+        MRC(MAKE_ ## _reg(11), ret); \
+        return ret; \
+    case 12: \
+        MRC(MAKE_ ## _reg(12), ret); \
+        return ret; \
+    case 13: \
+        MRC(MAKE_ ## _reg(13), ret); \
+        return ret; \
+    case 14: \
+        MRC(MAKE_ ## _reg(14), ret); \
+        return ret; \
+    case 15: \
+        MRC(MAKE_ ## _reg(15), ret); \
+        return ret; \
+    default: \
+        assert(bp_num == 0); \
+        MRC(MAKE_ ## _reg(0), ret); \
+        return ret; \
+    } \
+}
+
+/** Generates write functions for the CP14 control and value registers.
+ */
+#define DEBUG_GENERATE_WRITE_FN(_name, _reg)  \
+static void \
+_name(uint16_t bp_num, word_t val) \
+{ \
+    switch (bp_num) { \
+    case 1: \
+        MCR(MAKE_ ## _reg(1), val); \
+        return; \
+    case 2: \
+        MCR(MAKE_ ## _reg(2), val); \
+        return; \
+    case 3: \
+        MCR(MAKE_ ## _reg(3), val); \
+        return; \
+    case 4: \
+        MCR(MAKE_ ## _reg(4), val); \
+        return; \
+    case 5: \
+        MCR(MAKE_ ## _reg(5), val); \
+        return; \
+    case 6: \
+        MCR(MAKE_ ## _reg(6), val); \
+        return; \
+    case 7: \
+        MCR(MAKE_ ## _reg(7), val); \
+        return; \
+    case 8: \
+        MCR(MAKE_ ## _reg(8), val); \
+        return; \
+    case 9: \
+        MCR(MAKE_ ## _reg(9), val); \
+        return; \
+    case 10: \
+        MCR(MAKE_ ## _reg(10), val); \
+        return; \
+    case 11: \
+        MCR(MAKE_ ## _reg(11), val); \
+        return; \
+    case 12: \
+        MCR(MAKE_ ## _reg(12), val); \
+        return; \
+    case 13: \
+        MCR(MAKE_ ## _reg(13), val); \
+        return; \
+    case 14: \
+        MCR(MAKE_ ## _reg(14), val); \
+        return; \
+    case 15: \
+        MCR(MAKE_ ## _reg(15), val); \
+        return; \
+    default: \
+        assert(bp_num == 0); \
+        MCR(MAKE_ ## _reg(0), val); \
+        return; \
+    } \
+}
+
+#endif /* ARM_BASE_CP14_SAVE_AND_RESTORE */
